fixed some warning on windows VS with the latyes unit test added by Oliver Magy
also make the unit test optional since still has some link error in windows.

+++ newton-4.00/sdk/dCore/ndArray.h
@@ -234,7 +234,7 @@ void ndArray<T>::Resize(ndInt32 newSize)
 	if (newSize > m_capacity || (m_capacity == 0))
 	{
 		newSize = ndMax(newSize, 16);
-		T* const newArray = (T*)ndMemory::Malloc(ndInt32(sizeof(T) * newSize));
+		T* const newArray = (T*)ndMemory::Malloc(size_t(ndInt32(sizeof(T) * newSize)));
 		if (m_array) 
 		{
 			CopyData(newArray, m_array, m_size);
@@ -246,7 +246,7 @@ void ndArray<T>::Resize(ndInt32 newSize)
 	else if (newSize < m_capacity)
 	{
 		newSize = ndMax(newSize, 16);
-		T* const newArray = (T*)ndMemory::Malloc(ndInt32(sizeof(T) * newSize));
+		T* const newArray = (T*)ndMemory::Malloc(size_t(ndInt32(sizeof(T) * newSize)));
 		if (m_array) 
 		{
 			CopyData(newArray, m_array, newSize);
